Modificações nas Classes
