+++ app/Models/Disciplina.php
@@ -1,14 +1,17 @@
 <?php
 
-namespace App\Model;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use App\Models\Disciplina as Disciplina;
+
 
 class Disciplina extends Model
 {
     protected $fillable = ['nome', 'iniciais', 'cargaHoraria'];
+    public $timestamps = false;
 
     public function cursos(){
-        return $this->belongsToMany(Disciplina::class);
+        return $this->belongsToMany(Disciplina::class, 'cursos_disciplinas');
     }
 }
